+++ lib/services/safeNoteService.ts
@@ -1,34 +1,37 @@
 import safeNoteRepository from '../repositories/safeNoteRepository';
+import { SafeNoteCreate } from '../types/safeNoteTypes';
 import errorFactory from '../utils/errorFactory';
 
-async function createSafeNote(safeNoteData: { userId: any, label: any, content: any }){
+async function createSafeNote(userId: number, safeNoteData: SafeNoteCreate){
+
+    const safeNote = await safeNoteRepository.findByLabelAndUserId(userId, safeNoteData.label);
+    if(safeNote) throw errorFactory('conflict', 'There is already a safeNote registered with this label/title on your account.');
+
+    await safeNoteRepository.create(safeNoteData);
+
+}
+
+async function getById(userId: number, id: number){
+
+    const safeNote = await safeNoteRepository.getById(id);
+    if(!safeNote || safeNote.userId !== userId) throw errorFactory('not_found', 'Could not find the safeNote.');
+
+    return safeNote;
+
+}
+
+async function list(userId: number){
+
+    const safeNotes = await safeNoteRepository.list(userId);
+    return safeNotes
+
+}
+
+async function deleteSafeNote(userId: number, id: number){
+
+    const safeNote = await safeNoteRepository.getById(id);
+    if(!safeNote || safeNote.userId !== userId) throw errorFactory('not_found', 'Could not find the safeNote.');
     
-	 const { userId, label, content} = safeNoteData;
-    await safeNoteRepository.create({ userId, label, content });
-
-}
-
-async function getById(id: number){
-
-    await safeNoteRepository.getById(id);
-
-}
-
-async function list(){
-
-    await safeNoteRepository.list();
-
-}
-
-async function updateSafeNote(id: number, safeNoteData: { userId: any, label: any, content: any }){
-    
-	 const { userId, label, content } = safeNoteData;
-    await safeNoteRepository.update(id, { userId, label, content });
-
-}
-
-async function deleteSafeNote(id: number){
-
     await safeNoteRepository.deleteSafeNote(id);
 
 }
@@ -37,6 +40,5 @@ export default {
     createSafeNote,
     getById,
     list,
-    updateSafeNote,
     deleteSafeNote
 }
